officefloor/web/officeweb_executive/src/main/java/net/officefloor/web/executive/WebThreadAffinityExecutiveSource.java
/*
 * OfficeFloor - http://www.officefloor.net
 * Copyright (C) 2005-2018 Daniel Sagenschneider
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
package net.officefloor.web.executive;

import net.officefloor.frame.api.executive.Executive;
import net.officefloor.frame.api.executive.source.ExecutiveSource;
import net.officefloor.frame.api.executive.source.ExecutiveSourceContext;
import net.officefloor.frame.api.executive.source.impl.AbstractExecutiveSource;

/**
 * Web based {@link ExecutiveSource} providing {@link Thread} affinity.
 * 
 * @author Daniel Sagenschneider
 */
public class WebThreadAffinityExecutiveSource extends AbstractExecutiveSource {

	/*
	 * ==================== ExecutiveSource ==========================
	 */

	@Override
	protected void loadSpecification(SpecificationContext context) {
	}

	@Override
	public Executive createExecutive(ExecutiveSourceContext context) throws Exception {

		// Obtain the CPU cores
		CpuCore[] cpuCores = CpuCore.getCores();

		// Return the executive
		return new WebThreadAffinityExecutive(cpuCores, context);
	}

}